[depthMap] move camera struct in CUDA portion
