+++ src/aliceVision/depthMap/cuda/PlaneSweepingCuda.cpp
@@ -13,7 +13,6 @@
 #include <aliceVision/mvsUtils/common.hpp>
 #include <aliceVision/mvsUtils/fileIO.hpp>
 #include <aliceVision/depthMap/cuda/commonStructures.hpp>
-#include <aliceVision/depthMap/cuda/cameraStruct.hpp>
 #include <aliceVision/depthMap/cuda/planeSweeping/plane_sweeping_cuda.hpp>
 
 #include <iostream>
@@ -58,74 +57,7 @@ static void cps_fillCamera(cameraStruct* cam, int c, mvsUtils::MultiViewParams* 
     Matrix3x4 P = K * (mp->RArr[c] | (Point3d(0.0, 0.0, 0.0) - mp->RArr[c] * mp->CArr[c]));
     Matrix3x3 iP = mp->iRArr[c] * iK;
 
-    CameraBaseStruct& base = cam->cam;
-
-    base.C[0] = mp->CArr[c].x;
-    base.C[1] = mp->CArr[c].y;
-    base.C[2] = mp->CArr[c].z;
-
-    base.P[0] = P.m11;
-    base.P[1] = P.m21;
-    base.P[2] = P.m31;
-    base.P[3] = P.m12;
-    base.P[4] = P.m22;
-    base.P[5] = P.m32;
-    base.P[6] = P.m13;
-    base.P[7] = P.m23;
-    base.P[8] = P.m33;
-    base.P[9] = P.m14;
-    base.P[10] = P.m24;
-    base.P[11] = P.m34;
-
-    base.iP[0] = iP.m11;
-    base.iP[1] = iP.m21;
-    base.iP[2] = iP.m31;
-    base.iP[3] = iP.m12;
-    base.iP[4] = iP.m22;
-    base.iP[5] = iP.m32;
-    base.iP[6] = iP.m13;
-    base.iP[7] = iP.m23;
-    base.iP[8] = iP.m33;
-
-    base.R[0] = mp->RArr[c].m11;
-    base.R[1] = mp->RArr[c].m21;
-    base.R[2] = mp->RArr[c].m31;
-    base.R[3] = mp->RArr[c].m12;
-    base.R[4] = mp->RArr[c].m22;
-    base.R[5] = mp->RArr[c].m32;
-    base.R[6] = mp->RArr[c].m13;
-    base.R[7] = mp->RArr[c].m23;
-    base.R[8] = mp->RArr[c].m33;
-
-    base.iR[0] = mp->iRArr[c].m11;
-    base.iR[1] = mp->iRArr[c].m21;
-    base.iR[2] = mp->iRArr[c].m31;
-    base.iR[3] = mp->iRArr[c].m12;
-    base.iR[4] = mp->iRArr[c].m22;
-    base.iR[5] = mp->iRArr[c].m32;
-    base.iR[6] = mp->iRArr[c].m13;
-    base.iR[7] = mp->iRArr[c].m23;
-    base.iR[8] = mp->iRArr[c].m33;
-
-    base.K[0] = K.m11;
-    base.K[1] = K.m21;
-    base.K[2] = K.m31;
-    base.K[3] = K.m12;
-    base.K[4] = K.m22;
-    base.K[5] = K.m32;
-    base.K[6] = K.m13;
-    base.K[7] = K.m23;
-    base.K[8] = K.m33;
-
-    base.iK[0] = iK.m11;
-    base.iK[1] = iK.m21;
-    base.iK[2] = iK.m31;
-    base.iK[3] = iK.m12;
-    base.iK[4] = iK.m22;
-    base.iK[5] = iK.m32;
-    base.iK[6] = iK.m13;
-    base.iK[7] = iK.m23;
-    base.iK[8] = iK.m33;
+    cam->cam = ps_init_camera_vectors( mp->CArr[c], P, iP, mp->RArr[c], mp->iRArr[c], K, iK );
 
     if(cam->H != NULL)
     {
@@ -136,8 +68,6 @@ static void cps_fillCamera(cameraStruct* cam, int c, mvsUtils::MultiViewParams* 
     {
         cam->H = (*H);
     }
-
-    ps_init_camera_vectors( base );
 }
 
 static void cps_fillCameraData( mvsUtils::ImagesCache* ic, cameraStruct* cam, int c, const mvsUtils::MultiViewParams* mp)
@@ -228,7 +158,6 @@ PlaneSweepingCuda::PlaneSweepingCuda(int _CUDADeviceNo, mvsUtils::ImagesCache* _
     for(int rc = 0; rc < nImgsInGPUAtTime; ++rc)
     {
         _cams[rc] = new cameraStruct;
-        _cams[rc]->init();
         _camsRcs[rc] = -1;
         _camsTimes[rc] = clock();
     }
@@ -1283,7 +1212,6 @@ Point3d PlaneSweepingCuda::getDeviceMemoryInfo()
 //     for(int i = 0; i < rtcams->size(); i++)
 //     {
 //         cameraStruct* tcam = new cameraStruct;
-//         tcam->init();
 //         cps_fillCamera(tcam, (*rtcams)[i], mp, NULL, scale);
 //         ttcams[i] = tcam;
 //     }
@@ -1394,7 +1322,6 @@ Point3d PlaneSweepingCuda::getDeviceMemoryInfo()
 //     for(int i = 0; i < rtcams->size(); i++)
 //     {
 //         cameraStruct* tcam = new cameraStruct;
-//         tcam->init();
 //         cps_fillCamera(tcam, (*rtcams)[i], mp, NULL, scale);
 //         ttcams[i] = tcam;
 //     }
@@ -1480,7 +1407,6 @@ Point3d PlaneSweepingCuda::getDeviceMemoryInfo()
 //     for(int i = 0; i < rtcams->size(); i++)
 //     {
 //         cameraStruct* tcam = new cameraStruct;
-//         tcam->init();
 //         cps_fillCamera(tcam, (*rtcams)[i], mp, NULL, scale);
 //         ttcams[i] = tcam;
 //     }
@@ -1870,11 +1796,9 @@ bool PlaneSweepingCuda::optimizeDepthSimMapGradientDescent(StaticVector<DepthSim
 //     cameraStruct** ttcams = new cameraStruct*[2];
 //     {
 //         cameraStruct* tcam = new cameraStruct;
-//         tcam->init();
 //         cps_fillCamera(tcam, rc, mp, NULL, scale);
 //         ttcams[0] = tcam;
 //         tcam = new cameraStruct;
-//         tcam->init();
 //         cps_fillCamera(tcam, tc, mp, NULL, scale);
 //         ttcams[1] = tcam;
 //     }
